+++ src/bakalari_api/__main__.py
@@ -1,0 +1,5 @@
+"Demo."
+
+from bakalari_api.bakalari_demo import main
+
+main()