+++ src/bakalari_api/datastructure.py
@@ -1,11 +1,16 @@
 """Datastructure."""
+
 from __future__ import annotations
-from typing import Any
+
 from dataclasses import dataclass
+from typing import Any
 
 import orjson
 
-from const import Token
+from .const import Token
+from .logger_api import api_logger
+
+log = api_logger("Bakalari API").get()
 
 
 @dataclass
@@ -18,22 +23,27 @@ class Credentials:
     user_id: str = None
 
     @classmethod
-    def get(cls, data: dict[str, Any]) -> Credentials:
-        """ "Return class object."""
+    def create(cls, data: dict[str, Any]) -> Credentials:
+        """Create class object form data."""
+
         return cls(
             user_id=data[Token.USER_ID],
             access_token=data[Token.ACCESS_TOKEN],
             refresh_token=data[Token.REFRESH_TOKEN],
+            username=data[Token.USERNAME],
         )
+
     @classmethod
-    def get_from_json(cls, data: dict[str, Any]) -> Credentials:
+    def create_from_json(cls, data: dict[str, Any]) -> Credentials:
         """Return class object from JSON dictionary."""
-        return cls(
-            user_id=data["user_id"],
-            access_token=data["access_token"],
-            refresh_token=data["refresh_token"],
-            username=data["username"]
-        )   
+        return Credentials.create(
+            {
+                Token.USER_ID: data["user_id"],
+                Token.ACCESS_TOKEN: data["access_token"],
+                Token.REFRESH_TOKEN: data["refresh_token"],
+                Token.USERNAME: data["username"],
+            }
+        )
 
 
 @dataclass
@@ -49,48 +59,74 @@ class Schools:
     """List of schools with their url for Bakalari API."""
 
     def __init__(self) -> None:
+        """List of schools with their url for Bakalari API."""
+
         self.school_list: list[School] = []
 
     def append_school(
-        self, name: str = None, api_point: str = None, town: str = None
-    ) -> None:
-        """Append new school to the list."""
+        self,
+        name: str,
+        api_point: str,
+        town: str,
+    ) -> bool:
+        """Append new school to the school_list list."""
+
+        if name == "" or None:
+            log.error(
+                f"School's name should not be none or empty! (name={name}, api_point={api_point}, town={town})"
+            )
+            return False
+
+        if api_point == "" or None:
+            log.error(
+                f"School's API point should not be none or empty! (name={name}, api_point={api_point}, town={town})"
+            )
+            return False
+
+        if town == "" or None:
+            log.error(
+                f"School's town should not be none or empty! (name={name}, api_point={api_point}, town={town})"
+            )
+            return False
 
         new_school = School(name=name, api_point=api_point, town=town)
         self.school_list.append(new_school)
 
-    def get_url(self, name: str | None = None, idx: int | None = None) -> str | False:
-        """Returns url for school from name or index in dictionary.
-        Only one must be specified - name or index, otherwise returns False
+        return True
 
-        If name or index is not found in dictionary returns False."""
+    def get_url(self, name: str | None = None, idx: int | None = None) -> str | False:
+        """Return url for school from name or index in dictionary.
+
+        Only one must be specified - name or index, otherwise returns False
+        If name or index is not found in dictionary returns False.
+        """
 
         if (name is not None) and (idx is not None):
             return False
 
         if name is not None:
             for item in self.school_list:
-                if name in item.name:
+                if name == item.name:
                     return item.api_point
 
         if idx is not None:
-            return self.school_list[idx].api_point
+            try:
+                return self.school_list[idx].api_point
+            except IndexError:
+                return False
 
         return False
 
-    def get_schools_by_town(self, town: str = None) -> list[School]:
-        """Get list of schools in town.
-
-        Returns list of School obj."""
+    def get_schools_by_town(self, town: str | None = None) -> list[School]:
+        """Get list of schools in town."""
         _schools = []
         for item in self.school_list:
-            if town in item.town:
-                _schools.append(item)
+            _schools.append(item) if item.town in town else None
         return _schools
-    
+
     def get_school_name_by_api_point(self, api_point: str) -> str | bool:
         """Get school name by its api point."""
-        
+
         for item in self.school_list:
             if api_point == item.api_point:
                 return item.name
@@ -118,16 +154,17 @@ class Schools:
 
         return True
 
-    def load_from_file(self, filename: str) -> Schools:
-        """Loads schools list from a file.
-
-        Returns Schools obj."""
+    async def load_from_file(self, filename: str) -> Schools:
+        """Load schools list from a file."""
 
         try:
             with open(filename, mode="+rb") as file:
                 data = orjson.loads(file.read())
             file.close()
         except OSError:
+            return False
+        except orjson.JSONDecodeError:
+            log.error(f"Unable to decode JSON file. File {filename} is corrupted.")
             return False
 
         for item in data: